feat: Implement projection trait for global protocol types and restructure modules
- Added `implementations.rs` to define the `Project` trait implementations for various global protocol types, including `TChanSend`, `TChanRecv`, `TChanEnd`, `TChanStart`, `TChanChoice`, and `TChanPar`.
- Extracted helper traits for role-based dispatch and boolean logic into a new `helpers.rs` module.
- Updated `mod.rs` in the projection module to re-export essential helper types and removed embedded tests, moving them to a dedicated `tests.rs` file.
- Created a new `tests.rs` file to contain unit tests for the projection implementations, ensuring all tests pass successfully.
- Updated `TASKS.md` to reflect progress on code quality improvements and module restructuring tasks.
- Documented recent learnings and achievements in `learnings.md`, highlighting successful module restructuring and compliance with size guidelines.
