+++ work/TASKS.md
@@ -49,6 +49,48 @@ with a strong emphasis on the concepts outlined in `docs/duality.md`.
     - [x] **Task 1.1.5d**: Implement projection for `TChanPar`, `TChanRec`, `TChanEnd` constructs.
     - [x] **Task 1.1.5e**: Implement helper traits (`ProjectChoices`, `Not<Role>`, `ValidProjection`) for complex projections.
     - [x] **Task 1.1.5f**: Implement `ProjectionError` type and validation mechanisms. (**COMPLETED**)
+  - [ ] **Task 1.1.6**: Code Quality Improvements
+    - [x] **Task 1.1.6a**: Fix critical clippy warnings (empty line after doc comment, unused Sealed trait). (**COMPLETED** - 2025-05-24)
+    - [ ] **Task 1.1.6b**: Implement `Default` trait for all protocol types with `new()` methods to improve API ergonomics and satisfy clippy suggestions.
+      - [ ] Global Protocol Types (8): TChanSend, TChanRecv, TChanChoice, TChanOffer, TChanPar, TChanEnd, TChanStart  
+      - [ ] Local Protocol Types (6): EpChanSend, EpChanRecv, EpChanOffer, EpChanChoice, EpChanPar, EpChanEnd, EpChanStart
+      - **Benefit**: Allows `Default::default()` usage for cleaner API and better integration with generic code
+      - **Status**: 14 clippy suggestions available, auto-generated implementations look correct
+    - [ ] **Task 1.1.6c**: Module Structure Compliance (High Priority - guideline violations identified)
+      - [x] **Task 1.1.6c.1**: Extract embedded tests from duality and projection modules (immediate) ✅ COMPLETED
+        - [x] Extract 94-line test module from `duality/mod.rs` (lines 476-569) to `duality/tests.rs` ✅
+        - [x] Extract 69-line test module from `projection/mod.rs` (lines 460-528) to `projection/tests.rs` ✅
+        - [x] Update module declarations from embedded `mod tests {` to `mod tests;` ✅
+        - **Result**: Reduced duality from 569→476 lines, projection from 528→460 lines ✅
+      - [ ] **Task 1.1.6c.2**: Restructure oversized modules to meet 300-line guideline (advanced)
+        - [ ] Restructure `duality/` module (475 lines after test extraction)
+          - [ ] Extract helper traits to `duality/helpers.rs` (~50 lines)
+          - [ ] Extract global protocol implementations to `duality/global_impls.rs` (~150 lines)
+          - [ ] Extract local endpoint implementations to `duality/local_impls.rs` (~100 lines)
+          - [ ] Keep core trait and validation in `duality/mod.rs` (~175 lines)
+        - [ ] Restructure `projection/` module (459 lines after test extraction)
+          - [ ] Extract error types to `projection/errors.rs` (~50 lines)
+          - [ ] Extract validation logic to `projection/validation.rs` (~100 lines)
+          - [ ] Extract projection implementations to `projection/impls.rs` (~150 lines)
+          - [ ] Keep core trait in `projection/mod.rs` (~159 lines)
+        - [ ] Restructure `local/` module (448 lines)
+          - [ ] Extract endpoint type definitions to `local/endpoints.rs` (~200 lines)
+          - [ ] Extract trait implementations to `local/impls.rs` (~150 lines)
+          - [ ] Keep core types in `local/mod.rs` (~98 lines)
+        - [ ] Restructure `global/` module (434 lines)
+          - [ ] Extract protocol type definitions to `global/protocols.rs` (~200 lines)
+          - [ ] Extract trait implementations to `global/impls.rs` (~150 lines)
+          - [ ] Keep core types in `global/mod.rs` (~84 lines)
+      - [x] **Task 1.1.6c.3**: Validation and compliance verification ✅ COMPLETED
+        - [x] Verify all tests still pass: `cargo test` ✅ (35 tests pass)
+        - [x] Verify formatting: `cargo fmt --all -- --check` ✅
+        - [x] Verify linting: `cargo clippy` ✅
+        - [x] Verify compilation: `cargo build` ✅
+        - [x] Confirm test extraction successful ✅ (duality: 569→476 lines, projection: 528→460 lines)
+        - [ ] Confirm all modules now meet 300-line guideline (2 modules still exceed: local=448, global=434)
+        - [ ] Confirm implementation/test ratios meet 2:1 guideline (current: duality=5.1:1, projection=6.8:1)
+      - **Justification**: 4 out of 5 protocol modules exceed 300-line size guideline. Poor test organization with embedded tests in duality/projection modules. Non-compliant implementation/test ratios.
+      - **Priority**: High - affects maintainability and adherence to established coding standards
 - [ ] **Task 1.2**: Implement Label Preservation and Transformation Logic
   - [x] **Task 1.2.1**: Research label behavior in `Choice`, `Parallel`, `Rec` to inform design. (**COMPLETED** - documented in updated `duality.md`)
   - [x] **Task 1.2.2**: Design type-level traits for label transformations (e.g., `TMap`, `TCollect`, `TFilter`). (**COMPLETED** - detailed in updated `duality.md`)
